PXMB-37 PXMB-39 Implement POST, PUT, DELETE and GET by ID with proper validation

+++ src/main/java/com/startstepszalando/ecommerceshop/product/repository/ProductRepository.java
@@ -6,8 +6,10 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
+
 @Repository
 public interface ProductRepository extends JpaRepository<Product, Long> {
     @Query("SELECT CASE WHEN COUNT(p) > 0 THEN true ELSE false END FROM Product p WHERE p.name = :name AND p.price = :price")
-    boolean hasEntityWithNameAndPrice(@Param("name") String name, @Param("price") double price);
+    boolean hasEntityWithNameAndPrice(@Param("name") String name, @Param("price") BigDecimal price);
 }
